# swagger.json
{
    "swagger": "2.0",
     "info": {
        "description": "A Social platform for the creative at heart. Creates a community of like minded authors to foster inspiration and innovation by leveraging the modern web",
        "version": "1.0.0",
        "title": "Authors Haven",
        "contact": {
          "email": "noldor.ah@gmail.com"
        },
        "license": {
          "name": "MIT"
        }
      },
      "host": "noldor-ah-backend-staging.herokuapp.com",
      "basePath": "/api/v1",
...
      "schemes": [
        "https"
      ],
      "paths": {
...
        "/users/forgot/{hash}": {
          "post": {
            "tags": ["users"],
            "summary": "Resets the password with a provided hash",
            "description": "Provide your new password to complete reset password process",
            "operationId": "resetPassword",
            "consumes": ["application/x-www-form-urlencoded"],
            "produces" : ["application/json"],
            "parameters": [
            {
            "in": "path",
            "name": "hash",
            "type": "string",
              "required": true,
              "description": "Copy the hash just after the /forgot/ route"
            }, 
            {
                "name": "password",
                "in": "formData",
                "description": "Enter your password",
                "required": true,
                "type": "string"
              },
              {
                "name": "confirmPassword",
                "in": "formData",
                "description": "Enter your password",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "401": {
                "description": "invalid token"
              },
              "404": {
                "description": "Email not found"
              },
              "503": {
                "description": "password cannot be updated, try again"
              },
              "200": {
                "description": "password has been updated"
              }
            }
          }
        }
      }
  }